backend/requirements.txt
# Python 3.14 compatible (pydantic 2.12+ / pydantic-core 2.41+ have 3.14 wheels)
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database (MySQL)
sqlalchemy>=2.0.36
pymysql>=1.1.0
cryptography>=42.0.0

# Config and validation (2.12+ for Python 3.14)
pydantic>=2.12.0
pydantic-settings>=2.6.0
pydantic-core>=2.41.0

# File upload: PDF, Word, images (AI vision for PNG/JPEG)
python-multipart>=0.0.17
pdfplumber>=0.11.0
PyPDF2>=3.0.1
python-docx>=1.0.0

# OpenAI
openai>=1.55.0

# AWS S3 (resume uploads)
boto3>=1.35.0
